requirements.txt
Django==1.11.23
django-cleanup==3.2.0
django-registration-redux==2.6
Pillow==6.2.1
pytz==2019.1
sqlparse==0.3.0